compute_rsd_kernels/compute_rsd_biased_kernels return int

+++ include/biased_tracers.hpp
@@ -3,7 +3,7 @@
 
 class IntegrandTables;
 
-double compute_rsd_biased_kernels(
+int compute_rsd_biased_kernels(
         const int arguments[],
         int kernel_index,
         int n,
